backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]
sqlalchemy>=2.0
mysql-connector-python
pydantic>=2.0
requests
python-multipart
python-dotenv
